04_发票识别工具/pyproject.toml
[project]
name = "invoice-scanner"
version = "0.1.0"
description = "发票识别工具 - 扫描发票图片和PDF并导出到Excel"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pymupdf>=1.23.0",
    "pillow-heif>=0.15.0",
]
